\name{cci_bayesian_batch}
\alias{cci_bayesian_batch}
\title{Vectorised S4 batch.}
\description{Vectorised S4 batch.}
\usage{cci_bayesian_batch(dt, quan_map, cache, n_draws = 25L, alpha_0 = 10, seed = 42L, return_group_count = FALSE)}
\arguments{
  \item{dt}{Argument dt.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
  \item{n_draws}{Argument n_draws.}
  \item{alpha_0}{Argument alpha_0.}
  \item{seed}{Argument seed.}
  \item{return_group_count}{Argument return_group_count.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

