\name{cci_gold}
\alias{cci_gold}
\title{Compute the gold-standard Charlson Comorbidity Index for one encounter.}
\description{Compute the gold-standard Charlson Comorbidity Index for one encounter.}
\usage{cci_gold(icd_codes, quan_map)}
\arguments{
  \item{icd_codes}{Argument icd_codes.}
  \item{quan_map}{Argument quan_map.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

