\name{cci_gold_batch}
\alias{cci_gold_batch}
\title{Vectorised gold-standard CCI for a cohort.}
\description{Vectorised gold-standard CCI for a cohort.}
\usage{cci_gold_batch(dt, quan_map, pattern_lookup, dep_lookup)}
\arguments{
  \item{dt}{Argument dt.}
  \item{quan_map}{Argument quan_map.}
  \item{pattern_lookup}{Argument pattern_lookup.}
  \item{dep_lookup}{Argument dep_lookup.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

