\name{cci_interval}
\alias{cci_interval}
\title{Single-encounter S1 (interval) CCI estimator.}
\description{Single-encounter S1 (interval) CCI estimator.}
\usage{cci_interval(icd_anon, quan_map, cache)}
\arguments{
  \item{icd_anon}{Argument icd_anon.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

