\name{cci_interval_batch}
\alias{cci_interval_batch}
\title{Vectorised S1 batch.}
\description{Vectorised S1 batch.}
\usage{cci_interval_batch(dt, quan_map, cache)}
\arguments{
  \item{dt}{Argument dt.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

