\name{cci_mi_batch}
\alias{cci_mi_batch}
\title{Vectorised S3 batch.}
\description{Vectorised S3 batch.}
\usage{cci_mi_batch(dt, quan_map, cache, m = 20L, seed = 42L, return_group_count = FALSE)}
\arguments{
  \item{dt}{Argument dt.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
  \item{m}{Argument m.}
  \item{seed}{Argument seed.}
  \item{return_group_count}{Argument return_group_count.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

