\name{cci_probabilistic}
\alias{cci_probabilistic}
\title{Single-encounter S2 (probabilistic / inclusion-exclusion) CCI estimator.}
\description{Single-encounter S2 (probabilistic / inclusion-exclusion) CCI estimator.}
\usage{cci_probabilistic(icd_anon, quan_map, cache, age)}
\arguments{
  \item{icd_anon}{Argument icd_anon.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
  \item{age}{Argument age.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

