\name{cci_probabilistic_batch}
\alias{cci_probabilistic_batch}
\title{Vectorised S2 batch.}
\description{Vectorised S2 batch.}
\usage{cci_probabilistic_batch(dt, quan_map, cache, return_group_prob = FALSE)}
\arguments{
  \item{dt}{Argument dt.}
  \item{quan_map}{Argument quan_map.}
  \item{cache}{Argument cache.}
  \item{return_group_prob}{Argument return_group_prob.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

