\name{load_destatis}
\alias{load_destatis}
\title{Download and parse the Destatis 23131-01 reference frequency table.}
\description{Download and parse the Destatis 23131-01 reference frequency table.}
\usage{load_destatis(url = "https://www.destatis.de/static/DE/dokumente/5231301237015_SB.xlsx", cache = TRUE, force = FALSE, quiet = TRUE)}
\arguments{
  \item{url}{Argument url.}
  \item{cache}{Argument cache.}
  \item{force}{Argument force.}
  \item{quiet}{Argument quiet.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

