\name{qa_posterior_coverage}
\alias{qa_posterior_coverage}
\title{Share of encounters whose prediction is within k CCI points of gold.}
\description{Share of encounters whose prediction is within k CCI points of gold.}
\usage{qa_posterior_coverage(preds, strategies = c("s2_ecci", "s3_mi", "s4_bayes", "meta"))}
\arguments{
  \item{preds}{Argument preds.}
  \item{strategies}{Argument strategies.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

