\name{qa_score_distribution}
\alias{qa_score_distribution}
\title{CCI bin frequencies plus KS distance versus the gold distribution.}
\description{CCI bin frequencies plus KS distance versus the gold distribution.}
\usage{qa_score_distribution(preds, strategies = c("s2_ecci", "s3_mi", "s4_bayes", "meta"))}
\arguments{
  \item{preds}{Argument preds.}
  \item{strategies}{Argument strategies.}
}
\value{See package vignette and source for details.}
\examples{\dontrun{
# See README for usage examples.
}}

